+++ src/amocrm/amocrm.module.ts
@@ -1,7 +1,5 @@
 import { Module } from '@nestjs/common';
 import { ConfigService, ConfigModule } from '@nestjs/config';
-import { AmocrmConnector, AmocrmV2Auth } from './amocrm.connect';
-import { AmocrmV2Service, AmocrmV4Service } from './amocrm.service';
 import { HttpModule } from '@nestjs/axios';
 import { RedisModule } from '@app/redis/redis.module';
 import { LogModule } from '@app/log/log.module';
@@ -9,6 +7,10 @@ import { AsteriskCdrModule } from '@app/asterisk-cdr/asterisk-cdr.module';
 import { MongooseModule } from '@nestjs/mongoose';
 import { Amocrm, AmocrmSchema } from './amocrm.schema';
 import { getAmocrmV2Config, getAmocrmV4Config } from '@app/config/project-configs/amocrm.config';
+import { AmocrmV2Connector } from './v2/amocrm-v2.connect';
+import { AmocrmV4Connector } from './v4/amocrm-v4.connect';
+import { AmocrmV4Service } from './v4/amocrm-v4.service';
+import { AmocrmV2Service } from './v2/amocrm-v2.service';
 
 @Module({
   imports: [
@@ -29,11 +31,11 @@ import { getAmocrmV2Config, getAmocrmV4Config } from '@app/config/project-config
       useFactory: getAmocrmV4Config,
       inject: [ConfigService],
     },
-    AmocrmConnector,
+    AmocrmV2Connector,
+    AmocrmV4Connector,
     AmocrmV4Service,
     AmocrmV2Service,
-    AmocrmV2Auth,
   ],
-  exports: [AmocrmV4Service, AmocrmV2Service, AmocrmConnector],
+  exports: [AmocrmV4Service, AmocrmV2Service, AmocrmV4Connector],
 })
 export class AmocrmModule {}
