+++ src/amocrm/interfaces/amocrm.enum.ts
@@ -3,13 +3,13 @@ export enum ContactsOrder {
   id = 'id',
 }
 
-export enum AmoCRMAPIV2 {
+export enum AmocrmAPIV2 {
   auth = '/private/api/auth.php?type=json',
   account = '/private/api/v2/account',
   events = '/api/v2/events/',
 }
 
-export enum AmocrmAPI {
+export enum AmocrmAPIV4 {
   contacts = '/api/v4/contacts',
   leads = '/api/v4/leads',
   account = '/api/v4/account',
